Move back to using ViewModel to communicate between Fragment and Activity
